Update README.md
Stranger Things - In 1980s Indiana, a group of young friends witness supernatural forces and secret government exploits. As they search for answers, the children unravel a series of extraordinary mysteries.

The Conjuring - When famed paranormal investigators Ed and Lorraine Warren attempted to help them, a member of the family was possessed by an evil presence, and almost didn't live to tell the tale. These events later inspired a film adaptation, The Conjuring.

Saw X - Hoping for a miraculous cure, John Kramer travels to Mexico for a risky and experimental medical procedure, only to discover the entire operation is a scam to defraud the most vulnerable. Armed with a newfound purpose, the infamous serial killer uses deranged and ingenious traps to turn the tables on the con artists.